Refactor: created window module for gui-related code
